refactor: simplify component placement
