+++ src/app/(home)/_components/about.tsx
@@ -7,8 +7,8 @@ import {
   SplitSectionContent,
   SplitSectionSidebar,
 } from '@/components/layout/sections/split-section'
-import { ViewAnimation } from '@/components/shared/view-animation'
 import { buttonVariants } from '@/components/ui/button'
+import { ViewAnimation } from '@/components/view-animation'
 import { cn } from '@/lib/utils'
 
 export default function About(): React.ReactElement {
