+++ src/app/(home)/_components/contributions.tsx
@@ -13,12 +13,12 @@ import {
 } from '@/components/features/contributions/contribution-graph'
 import { Icons } from '@/components/icons/icons'
 import { Section } from '@/components/layout/sections/section'
-import { ViewAnimation } from '@/components/shared/view-animation'
 import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
 } from '@/components/ui/tooltip'
+import { ViewAnimation } from '@/components/view-animation'
 import { useIsMobile } from '@/hooks/use-mobile'
 import { cn } from '@/lib/utils'
