+++ src/app/(home)/_components/cta.tsx
@@ -2,8 +2,8 @@ import Link from 'next/link'
 import type React from 'react'
 import { Icons } from '@/components/icons/icons'
 import { Section } from '@/components/layout/sections/section'
-import { ViewAnimation } from '@/components/shared/view-animation'
 import { Button } from '@/components/ui/button'
+import { ViewAnimation } from '@/components/view-animation'
 
 export default function CTA(): React.ReactElement {
   return (
