+++ src/app/(home)/_components/hero.tsx
@@ -4,8 +4,8 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { Icons } from '@/components/icons/icons'
 import { Section } from '@/components/layout/sections/section'
-import { ViewAnimation } from '@/components/shared/view-animation'
 import { buttonVariants } from '@/components/ui/button'
+import { ViewAnimation } from '@/components/view-animation'
 import { linkItems } from '@/constants/navigation'
 import { baseOptions } from '@/constants/site'
 import { cn } from '@/lib/utils'
