+++ src/app/(home)/about/_components/detailed.tsx
@@ -6,8 +6,8 @@ import {
   SplitSectionHeader,
   SplitSectionSidebar,
 } from '@/components/layout/sections/split-section'
-import { ViewAnimation } from '@/components/shared/view-animation'
 import { buttonVariants } from '@/components/ui/button'
+import { ViewAnimation } from '@/components/view-animation'
 import { cn } from '@/lib/utils'
 
 export default function Detailed(): React.ReactElement {
