+++ src/app/(home)/about/_components/feed/index.tsx
@@ -4,7 +4,7 @@ import {
   SplitSectionHeader,
   SplitSectionSidebar,
 } from '@/components/layout/sections/split-section'
-import { ViewAnimation } from '@/components/shared/view-animation'
+import { ViewAnimation } from '@/components/view-animation'
 import { cn } from '@/lib/utils'
 import { getActivityItems } from './actions'
 import { GitHubEvent as GitHubEventComponent } from './event'
