+++ src/app/(home)/about/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
+import { ProfilePageJsonLd } from '@/components/json-ld'
 import { Wrapper } from '@/components/layout/wrapper'
-import { ProfilePageJsonLd } from '@/components/seo/json-ld'
 import { description as homeDescription } from '@/constants/site'
 import { createMetadata } from '@/lib/metadata'
 import { getSortedByDatePosts } from '@/lib/source'
