+++ src/app/(home)/blog/[slug]/page.tsx
@@ -7,8 +7,8 @@ import { GitHubCode } from '@/components/content/mdx/github-code'
 import { MdxContent } from '@/components/content/mdx/layout'
 import { LLMCopyButtonWithViewOptions } from '@/components/features/assistant/page-actions'
 import { TOCMinimap, TOCPopover } from '@/components/features/toc'
+import { PostJsonLd } from '@/components/json-ld'
 import { SectionBody } from '@/components/layout/sections/section-body'
-import { PostJsonLd } from '@/components/seo/json-ld'
 import { baseUrl } from '@/constants'
 import { owner, repo } from '@/constants/config/github'
 import { description as homeDescription } from '@/constants/site'
