+++ src/app/(home)/blog/page.tsx
@@ -1,11 +1,11 @@
 import type { Metadata, ResolvingMetadata } from 'next'
 import { notFound, redirect } from 'next/navigation'
 import Posts from '@/app/(home)/_components/posts'
+import { CollectionPageJsonLd } from '@/components/json-ld'
 import { Section } from '@/components/layout/sections/section'
 import { SectionBody } from '@/components/layout/sections/section-body'
 import { Wrapper } from '@/components/layout/wrapper'
-import { CollectionPageJsonLd } from '@/components/seo/json-ld'
-import { NumberedPagination } from '@/components/shared/numbered-pagination'
+import { NumberedPagination } from '@/components/numbered-pagination'
 import { postsPerPage } from '@/constants/config'
 import { createMetadata } from '@/lib/metadata'
 import { parsePageParam } from '@/lib/pagination'
