+++ src/app/(home)/blog/tags/[...slug]/page.tsx
@@ -2,12 +2,12 @@ import type { Metadata, ResolvingMetadata } from 'next'
 import { notFound, redirect } from 'next/navigation'
 import { PostCard } from '@/components/features/blog/post-card'
 import { Icons } from '@/components/icons/icons'
+import { TagJsonLd } from '@/components/json-ld'
 import { HeroSection } from '@/components/layout/sections/hero'
 import { Section } from '@/components/layout/sections/section'
 import { SectionBody } from '@/components/layout/sections/section-body'
-import { TagJsonLd } from '@/components/seo/json-ld'
-import { NumberedPagination } from '@/components/shared/numbered-pagination'
-import { ViewAnimation } from '@/components/shared/view-animation'
+import { NumberedPagination } from '@/components/numbered-pagination'
+import { ViewAnimation } from '@/components/view-animation'
 import { postsPerPage } from '@/constants/config'
 import { createMetadata } from '@/lib/metadata'
 import { parsePageParam } from '@/lib/pagination'
