+++ src/app/(home)/blog/tags/page.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from 'next'
 import { TagCard } from '@/components/features/blog/tag-card'
+import { CollectionPageJsonLd } from '@/components/json-ld'
 import { SectionBody } from '@/components/layout/sections/section-body'
 import { Wrapper } from '@/components/layout/wrapper'
-import { CollectionPageJsonLd } from '@/components/seo/json-ld'
-import { ViewAnimation } from '@/components/shared/view-animation'
+import { ViewAnimation } from '@/components/view-animation'
 import { title as homeTitle } from '@/constants/site'
 import { createMetadata } from '@/lib/metadata'
 import { getTags } from '@/lib/source'
