+++ src/app/(home)/uses/page.tsx
@@ -1,11 +1,11 @@
 import type { Metadata } from 'next'
 import type { ReactNode } from 'react'
 import { Fragment } from 'react'
+import { AboutPageJsonLd } from '@/components/json-ld'
 import { Section } from '@/components/layout/sections/section'
 import Separator from '@/components/layout/separator'
 import { Wrapper } from '@/components/layout/wrapper'
-import { AboutPageJsonLd } from '@/components/seo/json-ld'
-import { ViewAnimation } from '@/components/shared/view-animation'
+import { ViewAnimation } from '@/components/view-animation'
 import { hardware, software } from '@/constants/portfolio/uses'
 import { createMetadata } from '@/lib/metadata'
 import { Coding } from './_components/coding'
