+++ src/app/(home)/work/[slug]/page.tsx
@@ -2,16 +2,16 @@ import { TOCProvider } from 'fumadocs-ui/components/toc'
 import type { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 import { type ReactElement, ViewTransition } from 'react'
+import { BlurImage } from '@/components/blur-image'
 import { mdxComponents } from '@/components/content/mdx/components'
 import { GitHubCode } from '@/components/content/mdx/github-code'
 import { PreviewButton } from '@/components/content/mdx/preview-button'
 import { TOCMinimap, TOCPopover } from '@/components/features/toc'
 import { Icons } from '@/components/icons/icons'
+import { WorkJsonLd } from '@/components/json-ld'
 import { Section } from '@/components/layout/sections/section'
 import { SectionBody } from '@/components/layout/sections/section-body'
-import { WorkJsonLd } from '@/components/seo/json-ld'
-import { BlurImage } from '@/components/shared/blur-image'
-import { ViewAnimation } from '@/components/shared/view-animation'
+import { ViewAnimation } from '@/components/view-animation'
 import { description as homeDescription } from '@/constants/site'
 import { createMetadata, getWorkPageImage } from '@/lib/metadata'
 import { getWork, getWorkPages, type WorkPage as MDXPage } from '@/lib/source'
