+++ src/app/(home)/work/page.tsx
@@ -1,12 +1,12 @@
 import type { Metadata, ResolvingMetadata } from 'next'
 import { notFound, redirect } from 'next/navigation'
 import { WorkCard } from '@/components/features/work/work-card'
+import { CollectionPageJsonLd } from '@/components/json-ld'
 import { Section } from '@/components/layout/sections/section'
 import { Wrapper } from '@/components/layout/wrapper'
-import { CollectionPageJsonLd } from '@/components/seo/json-ld'
-import { NumberedPagination } from '@/components/shared/numbered-pagination'
-import { SearchRedirectInput } from '@/components/shared/search-redirect-input'
-import { ViewAnimation } from '@/components/shared/view-animation'
+import { NumberedPagination } from '@/components/numbered-pagination'
+import { SearchRedirectInput } from '@/components/search-redirect-input'
+import { ViewAnimation } from '@/components/view-animation'
 import { worksPerPage } from '@/constants/config'
 import { createMetadata } from '@/lib/metadata'
 import { parsePageParam } from '@/lib/pagination'
