+++ src/app/provider.tsx
@@ -4,10 +4,10 @@ import { ProgressProvider } from '@bprogress/next/app'
 import { SoundProvider } from '@web-kits/audio/react'
 import { NuqsAdapter } from 'nuqs/adapters/next/app'
 import type { ReactNode } from 'react'
+import Analytics from '@/components/analytics'
 import { AISearch } from '@/components/features/assistant/chat'
 import { AISearchTrigger } from '@/components/features/assistant/search-trigger'
 import { MobileNav } from '@/components/layout/header/mobile/mobile-nav'
-import Analytics from '@/components/providers/analytics'
 import { SmoothCursor } from '@/components/providers/smooth-cursor'
 import { TailwindIndicator } from '@/components/providers/tailwind-indicator'
 import { Toaster } from '@/components/ui/sonner'
