+++ src/components/blur-image.tsx
