+++ src/components/content/mdx/link-preview.client.tsx
@@ -2,8 +2,8 @@
 
 import Image from 'next/image'
 import type { ReactNode } from 'react'
+import { BlurImage } from '@/components/blur-image'
 import { Icons } from '@/components/icons/icons'
-import { BlurImage } from '@/components/shared/blur-image'
 import {
   HoverCard,
   HoverCardContent,
