+++ src/components/content/mdx/preview-button.tsx
@@ -1,8 +1,8 @@
 import Image from 'next/image'
 import Link from 'next/link'
 import type { ReactNode } from 'react'
+import { BlurImage } from '@/components/blur-image'
 import { Icons } from '@/components/icons/icons'
-import { BlurImage } from '@/components/shared/blur-image'
 import { Button } from '@/components/ui/button'
 import {
   HoverCard,
