+++ src/components/features/work/work-card.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { ViewTransition } from 'react'
+import { BlurImage } from '@/components/blur-image'
 import { Icons } from '@/components/icons/icons'
-import { BlurImage } from '@/components/shared/blur-image'
 
 interface WorkCardProps {
   description: string
