+++ src/components/layout/footer/links.tsx
@@ -2,7 +2,7 @@ import { resolveLinkItems } from 'fumadocs-ui/layouts/shared'
 import Link from 'next/link'
 import type { ReactNode } from 'react'
 import { ActiveLink } from '@/components/layout/footer/active-link'
-import { ViewAnimation } from '@/components/shared/view-animation'
+import { ViewAnimation } from '@/components/view-animation'
 import { linkItems, socials } from '@/constants/navigation'
 import { baseOptions } from '@/constants/site'
 import { getSortedWork } from '@/lib/source'
