+++ src/components/layout/sections/section.tsx
@@ -1,6 +1,6 @@
 import type { HTMLAttributes } from 'react'
 import { Icons } from '@/components/icons/icons'
-import { ViewAnimation } from '@/components/shared/view-animation'
+import { ViewAnimation } from '@/components/view-animation'
 import { cn } from '@/lib/utils'
 
 type SectionProps = {
