+++ src/components/search-redirect-input.tsx
